tests/requirements.txt
pytest==3.0.0
pytest-django
mock==2.0.0
freezegun==0.3.7
